backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# AI Integration
google-generativeai==0.3.2

# Docker Integration
docker==7.0.0

# Proxmox Integration
proxmoxer==2.0.1
requests==2.31.0

# Utilities
python-dateutil==2.8.2

# Web Parsing
beautifulsoup4==4.12.3
